requirements.txt
torch
numpy
matplotlib
scipy.stats
random
pandas
scikit-learn